+++ poly/src/polynomial/mod.rs
@@ -60,15 +60,15 @@ pub trait DenseMVPolynomial<F: Field>: Polynomial<F> {
     /// The type of the terms of `self`
     type Term: multivariate::Term;
 
-    /// Constructs a new polynomial from a list of tuples of the form `(Self::Term, coeff)`
+    /// Constructs a new polynomial from a list of tuples of the form `(coeff, Self::Term)`
     fn from_coefficients_slice(num_vars: usize, terms: &[(F, Self::Term)]) -> Self {
         Self::from_coefficients_vec(num_vars, terms.to_vec())
     }
 
-    /// Constructs a new polynomial from a list of tuples of the form `(Self::Term, coeff)`
+    /// Constructs a new polynomial from a list of tuples of the form `(coeff, Self::Term)`
     fn from_coefficients_vec(num_vars: usize, terms: Vec<(F, Self::Term)>) -> Self;
 
-    /// Returns the terms of a `self` as a list of tuples of the form `(Self::Term, coeff)`
+    /// Returns the terms of a `self` as a list of tuples of the form `(coeff, Self::Term)`
     fn terms(&self) -> &[(F, Self::Term)];
 
     /// Returns the number of variables in `self`