+++ poly/src/polynomial/multivariate/sparse.rs
@@ -38,6 +38,21 @@ impl<F: Field> Polynomial<F> for SparsePolynomial<F, SparseTerm> {
     type Point = Vec<F>;
 
     /// Returns the total degree of the polynomial
+    ///
+    /// # Examples
+    /// ```
+    /// use ark_poly::{
+    ///     polynomial::multivariate::{SparsePolynomial, SparseTerm},
+    ///     DenseMVPolynomial, Polynomial,
+    /// };
+    /// use ark_std::test_rng;
+    /// use ark_test_curves::bls12_381::Fq;
+    ///
+    /// let rng = &mut test_rng();
+    /// // Create a multivariate polynomial of degree 7
+    /// let poly: SparsePolynomial<Fq, SparseTerm> = SparsePolynomial::rand(7, 2, rng);
+    /// assert_eq!(poly.degree(), 7);
+    /// ```
     fn degree(&self) -> usize {
         self.terms
             .iter()
@@ -47,6 +62,23 @@ impl<F: Field> Polynomial<F> for SparsePolynomial<F, SparseTerm> {
     }
 
     /// Evaluates `self` at the given `point` in `Self::Point`.
+    ///
+    /// # Examples
+    /// ```
+    /// use ark_poly::{
+    ///     polynomial::multivariate::{SparsePolynomial, SparseTerm, Term},
+    ///     DenseMVPolynomial, Polynomial,
+    /// };
+    /// use ark_ff::UniformRand;
+    /// use ark_std::test_rng;
+    /// use ark_test_curves::bls12_381::Fq;
+    ///
+    /// let rng = &mut test_rng();
+    /// let poly = SparsePolynomial::rand(4, 3, rng);
+    /// let random_point = vec![Fq::rand(rng); 3];
+    /// // The result will be a single element in the field.
+    /// let result: Fq = poly.evaluate(&random_point);
+    /// ```
     fn evaluate(&self, point: &Vec<F>) -> F {
         assert!(point.len() >= self.num_vars, "Invalid evaluation domain");
         if self.is_zero() {
@@ -80,6 +112,27 @@ impl<F: Field> DenseMVPolynomial<F> for SparsePolynomial<F, SparseTerm> {
     type Term = SparseTerm;
 
     /// Constructs a new polynomial from a list of tuples of the form `(coeff, Self::Term)`
+    ///
+    /// # Examples
+    /// ```
+    /// use ark_poly::{
+    ///     polynomial::multivariate::{SparsePolynomial, SparseTerm, Term},
+    ///     DenseMVPolynomial, Polynomial,
+    /// };
+    /// use ark_test_curves::bls12_381::Fq;
+    ///
+    /// // Create a multivariate polynomial in 3 variables, with 4 terms:
+    /// // 2*x_0^3 + x_0*x_2 + x_1*x_2 + 5
+    /// let poly = SparsePolynomial::from_coefficients_vec(
+    ///     3,
+    ///     vec![
+    ///         (Fq::from(2), SparseTerm::new(vec![(0, 3)])),
+    ///         (Fq::from(1), SparseTerm::new(vec![(0, 1), (2, 1)])),
+    ///         (Fq::from(1), SparseTerm::new(vec![(1, 1), (2, 1)])),
+    ///         (Fq::from(5), SparseTerm::new(vec![])),
+    ///     ],
+    /// );
+    /// ```
     fn from_coefficients_vec(num_vars: usize, mut terms: Vec<(F, SparseTerm)>) -> Self {
         // Ensure that terms are in ascending order.
         terms.sort_by(|(_, t1), (_, t2)| t1.cmp(t2));